app-misc/rmlint/

+++ app-misc/rmlint/rmlint-2.10.3.ebuild
@@ -99,13 +99,13 @@ src_install() {
 	escons "${scons_opts[@]}" --prefix="${ED}/usr" --actual-prefix="${EPREFIX}/usr" install
 
 	# https://github.com/sahib/rmlint/pull/525
-	#if use doc; then
-	#	gzip -d "${ED}/usr/share/man/man1/rmlint.1.gz" || die
-	#fi
+	if use doc; then
+		gzip -d "${ED}/usr/share/man/man1/rmlint.1.gz" || die
+	fi
 	if use gui; then
 		python_optimize
 	fi
-	#einstalldocs
+	einstalldocs
 	find . -type f -name gschemas.compiled -delete
 }
